Minor renaming, start of scaladocs
